Make reusable components for use in new page
